requirements.txt
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dotenv==1.0.0
python-telegram-bot==20.5
redis==5.0.0
sniffio==1.3.0
typing_extensions==4.7.1
